Chore: 하나의 js 파일에 뭉쳐있던 함수들을 작은 파일들로 분리
헤더의 연월을 표시하는 부분의 기능을 담당하는 파일에 있던 여러 함수들을
작은 파일들로 나누어 모듈화 하였다.

그리고 담당하는 기능별로 파일들을 잘 분류하는 작업을 하고 있는데 어떻게
해야 잘 나눌 수 있을까 아직 고민중이다.
