+++ src/js/components/header/initHeaderMonthYear.js
@@ -1,0 +1,7 @@
+import { displayCurrentMonthYear } from './displayCurrentMonthYear.js';
+import { checkBtnClickEvent } from './checkBtnClickEvent.js';
+
+export function initHeaderMonthYear() {
+	displayCurrentMonthYear();
+	checkBtnClickEvent();
+}
