+++ src/js/utils/submitBtnHandelr.js
@@ -1,17 +1,3 @@
-function formEventChecker() {
-	const form = document.querySelector('form');
-	form.addEventListener('input', btnControl);
-	form.addEventListener('click', btnControl);
-}
-
-function btnControl() {
-	if (isAllFilled()) {
-		activateBtn();
-	} else {
-		deactivateBtn();
-	}
-}
-
 function isAllFilled() {
 	const inputs = document.querySelectorAll('form input[type=text]');
 	for (const input of inputs) {
@@ -30,6 +16,20 @@ function deactivateBtn() {
 	const submitBtn = document.querySelector('.submit-btn svg');
 	submitBtn.children[0].setAttribute('fill', '#A79FCB');
 	submitBtn.children[0].setAttribute('fill-opacity', '.4');
+}
+
+function btnControl() {
+	if (isAllFilled()) {
+		activateBtn();
+	} else {
+		deactivateBtn();
+	}
+}
+
+function formEventChecker() {
+	const form = document.querySelector('form');
+	form.addEventListener('input', btnControl);
+	form.addEventListener('click', btnControl);
 }
 
 formEventChecker();
